requirements.txt
flask
flask-bootstrap
tensorflow
pillow
